likeService - post.save 삭제

+++ src/main/java/com/sparta/springweb/model/Posts.java
@@ -51,7 +51,7 @@ public class Posts extends Timestamped {
     private User user;
 
     @Transient
-    private long likesCount;
+    private int likesCount;
 
     @Transient
     private boolean likesState;
